expand *fix_bounds algo FIXME, +verify method
"Refactored" the fix_bounds algorithm. It is hard to say if it works the
same way, especially as it ran into an unreported error (process was
simply killed). The new versions runs through and works as the original
intention.

The verify() routine reveals that there is still a logic bug within the
algo, as it does not lead to the desired boundary condition.

FIXME - find logic bug and verify again.

On branch feature/expand
Changes to be committed:
      modified:   estss/expand.py