feat *added fs=1000 argument to tsfel calc to prevent bug
tsfel has a bug in the fft calculation that rounds the freq/2 endpoint
to the nearest integer, with the previous fs=1 (read from the csv file),
this will generate an all-zero frequency vector, leading to
miscalculated frequency features. By specifying fs=1000 as a global in
the function call (overwrites the individual fs from csv file), this
vectzor is calculated correctly and dependend features as well (some may
be off by a factor of 1000, but this scaling does not matter due to
later normalization).

On branch feature/features

Changes to be committed:
      modified:   estss/features.py