reduce +first version of reduce_chain/compute_reduced_sets
First version of the whole compute chain that automatically loads the
default dataframes and computes the n2048...n32 sets.

features.py module now has a get_features and compute_features function
to be in line with the init and expand module

reduce.py module now has a get_reduced_sets and compute_reduced_sets
function to be in line with the init, expand and feature module

On branch feature/reduce
Your branch is up to date with 'origin/feature/reduce'.

Changes to be committed:
      modified:   estss/expand.py
      modified:   estss/features.py
      modified:   estss/reduce.py