custom +features calculation in chunks, works too slow
Feature calculation is done in 8 chunks of manifold ts to stay within
RAM restrictions. However, feature calculation for a single ts with 8760
points needs ~4s instead of ~0.15s for a ts with 1000 points (summing up
to 25 days instead of 10-15 hours for 500 000 ts). Following,
calculation must be moved to cluster.

On branch feature/custom
Changes to be committed:
      modified:   examples/redispatch/features_rd.py